It seems like $this->today->year...this->birth->year can also be of type double. However, the property $age is declared as type integer. Added an additional type check

+++ Utility/AgeCalculator.php
@@ -88,6 +88,6 @@ class AgeCalculator
 
     private function setAgeByYear()
     {
-        $this->age = $this->today->year - $this->birth->year;
+        $this->age = (int) ($this->today->year - $this->birth->year);
     }
 }
